Remove `Of::weekday`
